Release 1.03.14
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@658 98333e67-4a24-44d7-a75c-e53540dd3050

+++ branches/release_1_03/xmlrpc-c/src/xmlrpc_server_abyss.c
@@ -126,9 +126,9 @@ traceChunkRead(TSession * const abyssSessionP) {
 
 static void
 get_buffer_data(TSession * const r, 
-                int        const max, 
+                size_t     const max, 
                 char **    const out_start, 
-                int *      const out_len) {
+                size_t *   const out_len) {
 
     /* Point to the start of our data. */
     *out_start = &r->conn->buffer[r->conn->bufferpos];
